Refactored the ProteinModificationsTable
