+++ psi-mod-CV-browser/src/components.d.ts
@@ -14,7 +14,8 @@ declare module 'vue' {
     AppFooter: typeof import('./components/AppFooter.vue')['default']
     HelloWorld: typeof import('./components/HelloWorld.vue')['default']
     ProteinModificationDetailsDialog: typeof import('./components/modifications/ProteinModificationDetailsDialog.vue')['default']
-    ProteinModificationsFilters: typeof import('./components/ProteinModificationsFilters.vue')['default']
+    ProteinModificationsDataTable: typeof import('./components/table/ProteinModificationsDataTable.vue')['default']
+    ProteinModificationsFilters: typeof import('./components/table/ProteinModificationsFilters.vue')['default']
     ProteinModificationsMainTable: typeof import('./components/ProteinModificationsMainTable.vue')['default']
     ProteinModificationsTable: typeof import('./components/table/ProteinModificationsTable.vue')['default']
     RouterLink: typeof import('vue-router')['RouterLink']
